fix: updating the board context and adapting the components accordingly
move `uniqueTagsFromTasks` memo in context to avoid code repeat
